refactor: moving access_result to shared_kernel
