+++ server/src/password_management_context/application/use_cases/update_password_use_case.py
@@ -2,9 +2,8 @@ from password_management_context.application.gateways import PasswordRepository
 from password_management_context.domain.entities import Password
 from password_management_context.application.commands import UpdatePasswordCommand
 from password_management_context.domain.exceptions import PasswordNotFoundError
-from rights_access_context.application.responses.access_result import Granted
+from shared_kernel.access_control import Granted, AccessController, AccessDeniedError
 from shared_kernel.encryption import EncryptionService
-from shared_kernel.access_control import AccessController, AccessDeniedError
 
 
 class UpdatePasswordUseCase:
