--- server/src/rights_access_context/application/responses/__init__.py
@@ -1,4 +1,0 @@
-from .access_result import AccessResult
-from .access_result import Granted
-
-__all__ = ["AccessResult", "Granted"]
