+++ server/src/rights_access_context/application/use_cases/check_access_use_case.py
@@ -3,7 +3,7 @@ from uuid import UUID
 from rights_access_context.application.gateways import (
     RightsRepository,
 )
-from rights_access_context.application.responses import AccessResult, Granted
+from shared_kernel.access_control import AccessResult, Granted
 from rights_access_context.domain.value_objects import Permission
