+++ server/src/shared_kernel/access_control/__init__.py
@@ -1,5 +1,6 @@
+from .access_result import AccessResult, Granted
 from .access_controller import AccessController
 from .exceptions import AccessDeniedError
 
 # AccessController provides unified access control (both check and grant operations)
-__all__ = ["AccessController", "AccessDeniedError"]
+__all__ = ["AccessController", "AccessDeniedError", "AccessResult", "Granted"]
